Organize App files modularly
